Making an unconventional one-time commit directly to master in order to keep this repo up to date with what has actually currently been deployed to production in the last couple months.
Some of the changes included in this commit may appear as entirely changed files, but that would be the effect of Windows-style carriage returns getting added to the end of each line (CRLF line endings instead of LF).  For future reference, 'diff -bw' can be run to ignore white-space differences when diffing, and for 'git diff', the '--ignore-space-at-eol' flag can be added to just show the salient differences and ignore the end-of-line characters which show up as '^M' in 'git diff'.

The changes included in this commit include:
- Adding Google Analytics code to the search dropdown to track what dropdown option was chosen, if any, when users click 'Go' or the magnifying glass.  This addresses #133 while side-stepping issues raised there about code duplication and alt attributes on <input>s.
- Adding Eckles's Instagram to the footer. Fixes #132.
- Adding changes recently requested for the UTLC's navigation menu. Fixes #131.
- Some random new line in templates/html.tpl.php that happens to be on production.